backend/requirements.txt
crewai==0.11.0
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0